+++ sdk/identity/azure-identity/azure/identity/_internal/__init__.py
@@ -36,7 +36,7 @@ from .auth_code_redirect_handler import AuthCodeRedirectServer
 from .aadclient_certificate import AadClientCertificate
 from .certificate_credential_base import CertificateCredentialBase
 from .client_secret_credential_base import ClientSecretCredentialBase
-from .exception_wrapper import wrap_exceptions
+from .decorators import wrap_exceptions
 from .msal_credentials import InteractiveCredential
