+++ sdk/identity/azure-identity/azure/identity/_internal/decorators.py
@@ -5,6 +5,9 @@
 import functools
 import logging
 
+from six import raise_from
+from azure.core.exceptions import ClientAuthenticationError
+
 
 def log_get_token(logger, class_name):
     """Adds logging around get_token calls.
@@ -30,3 +33,19 @@ def log_get_token(logger, class_name):
         return wrapper
 
     return decorator
+
+
+def wrap_exceptions(fn):
+    """Prevents leaking exceptions defined outside azure-core by raising ClientAuthenticationError from them."""
+
+    @functools.wraps(fn)
+    def wrapper(*args, **kwargs):
+        try:
+            return fn(*args, **kwargs)
+        except ClientAuthenticationError:
+            raise
+        except Exception as ex:  # pylint:disable=broad-except
+            auth_error = ClientAuthenticationError(message="Authentication failed: {}".format(ex))
+            raise_from(auth_error, ex)
+
+    return wrapper
