+++ sdk/identity/azure-identity/azure/identity/aio/_internal/__init__.py
@@ -3,6 +3,6 @@
 # Licensed under the MIT License.
 # ------------------------------------
 from .aad_client import AadClient
-from .exception_wrapper import wrap_exceptions
+from .decorators import wrap_exceptions
 
 __all__ = ["AadClient", "wrap_exceptions"]
