+++ sdk/identity/azure-identity/azure/identity/aio/_internal/decorators.py
@@ -5,6 +5,8 @@
 import functools
 import logging
 
+from azure.core.exceptions import ClientAuthenticationError
+
 
 def log_get_token(logger):
     """Adds logging around get_token calls.
@@ -26,3 +28,20 @@ def log_get_token(logger):
         return wrapper
 
     return decorator
+
+
+def wrap_exceptions(fn):
+    """Prevents leaking exceptions defined outside azure-core by raising ClientAuthenticationError from them."""
+
+    @functools.wraps(fn)
+    async def wrapper(*args, **kwargs):
+        try:
+            result = await fn(*args, **kwargs)
+            return result
+        except ClientAuthenticationError:
+            raise
+        except Exception as ex:  # pylint:disable=broad-except
+            auth_error = ClientAuthenticationError(message="Authentication failed: {}".format(ex))
+            raise auth_error from ex
+
+    return wrapper
